Aha - when I try the --with-arithlib build on the Mac it fails to compile because Xcode/clang does not view int64_t as enough like intptr_t on a 64-bit machine, so it flags a whole pile of moans about overloading that it things is ambiguous. Well when I am feeling calm I can view this level of being picky as desirable - eg the same source code might be closer to ambiguous on a 32-bit system. But providing overloaded functions that accept int, int32_t, ssize_t, intptr_t, int64_t without some of those looking like repetition or others making the resolution of overloading ambiguous is something I have not mastered yet! I will need to clean my code up somewhat to make it build properly. But now that is going tgo be next week. I believe that version without arithlib should be OK!
git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@7405 2bfe0521-f11c-4a00-b80e-6202646ff360

+++ csl/cslbase/arith-plus.cpp
@@ -100,7 +100,7 @@ LispObject Plus::op(double a, Fixnum b)
 
 // long float + fixnum
 LispObject Plus::op(FLOAT_128 a, Fixnum b)
-{   return make_boxfloat128(a + (FLOAT_128)int_of_fixnum(b));
+{   return make_boxfloat128(a + (FLOAT_128)(int64_t)int_of_fixnum(b));
 }
 
 // fixnum + bignum
@@ -439,7 +439,7 @@ LispObject Difference::op(double a, Fixnum b)
 
 // long float - fixnum
 LispObject Difference::op(FLOAT_128 a, Fixnum b)
-{   return make_boxfloat128(a - (FLOAT_128)(int_of_fixnum(b)));
+{   return make_boxfloat128(a - (FLOAT_128)(int64_t)int_of_fixnum(b));
 }
 
 // fixnum - bignum
@@ -715,7 +715,7 @@ LispObject Difference::op(FLOAT_128 a, double b)
 
 // fixnum - long float
 LispObject Difference::op(Fixnum a, FLOAT_128 b)
-{   return make_boxfloat128((FLOAT_128)int_of_fixnum(a) - b);
+{   return make_boxfloat128((FLOAT_128)(int64_t)int_of_fixnum(a) - b);
 }
 
 // bignum - long float
